Refactor : 맵을 나타내는 좌표 데이터들 따로 관리
